requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
networkx>=3.3
pydantic>=2.0
google-genai>=1.0
